[fix-37133][logminer] 解决日志文件过多 导致查询超时

+++ flinkx-oraclelogminer/flinkx-oraclelogminer-reader/src/main/java/com/dtstack/flinkx/oraclelogminer/entity/QueueData.java
@@ -17,6 +17,7 @@
  */
 package com.dtstack.flinkx.oraclelogminer.entity;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 /**
@@ -26,15 +27,15 @@ import java.util.Map;
  * @author tudou
  */
 public class QueueData {
-    private long scn;
+    private BigDecimal scn;
     private Map<String, Object> data;
 
-    public QueueData(long lsn, Map<String, Object> data) {
+    public QueueData(BigDecimal lsn, Map<String, Object> data) {
         this.scn = lsn;
         this.data = data;
     }
 
-    public long getScn() {
+    public BigDecimal getScn() {
         return scn;
     }
